# apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Football players API ",
    "description": "List of football players with their details",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "tags": [
    {
      "name": "Players",
      "description": "API for football players in the system"
    }
  ],
  "paths": {
    "/players": {
      "get": {
        "tags": ["Players"],
        "summary": "Get all players",
        "description": "Get all players in the system",
        "operationId": "getPlayers",
        "responses": {
          "200": {
            "description": "A list of players.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Players"],
        "summary": "Add a new player",
        "description": "Add a new player to the system",
        "operationId": "addPlayer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Player added"
          }
        }
      }
    },
    "/players/{id}": {
      "get": {
        "tags": ["Players"],
        "summary": "Get player by id",
        "description": "Get a player by its id",
        "operationId": "getPlayerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of player to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A player.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "404": {
            "description": "Player not found"
          }
        }
      },
      "put": {
        "tags": ["Players"],
        "summary": "Update an existing player",
        "description": "Update an existing player in the system",
        "operationId": "updatePlayer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of player to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Player updated"
          },
          "404": {
            "description": "Player not found"
          }
        }
      },
      "delete": {
        "tags": ["Players"],
        "summary": "Delete a player",
        "description": "Delete a player in the system",
        "operationId": "deletePlayer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of player to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player deleted"
          },
          "404": {
            "description": "Player not found"
          }
        }
      }
    }
  }
}